+++ src/TRegx/CleanRegex/ForArray/ForArrayPatternImpl.php
@@ -6,29 +6,37 @@ use TRegx\CleanRegex\Internal\InternalPattern;
 
 class ForArrayPatternImpl implements ForArrayPattern
 {
-    /** @var array */
-    private $array;
     /** @var InternalPattern */
     private $pattern;
+    /** @var array */
+    private $array;
+    /** @var bool */
+    private $throwOnNonStringElements;
 
-    public function __construct(InternalPattern $pattern, array $array)
+    public function __construct(InternalPattern $pattern, array $array, bool $strict = false)
     {
         $this->pattern = $pattern;
         $this->array = $array;
+        $this->throwOnNonStringElements = $strict;
     }
 
     public function filter(): array
     {
-        return (new FilterArrayPattern($this->pattern, $this->array))->filter();
+        return (new FilterArrayPattern($this->pattern, $this->array, $this->throwOnNonStringElements))->filter();
     }
 
     public function filterAssoc(): array
     {
-        return (new FilterArrayPattern($this->pattern, $this->array))->filterAssoc();
+        return (new FilterArrayPattern($this->pattern, $this->array, $this->throwOnNonStringElements))->filterAssoc();
     }
 
     public function filterByKeys(): array
     {
         return (new FilterArrayKeysPattern($this->pattern, $this->array))->filterByKeys(new PregGrepArrayIntersectStrategy());
+    }
+
+    public function strict(): ForArrayPattern
+    {
+        return new self($this->pattern, $this->array, true);
     }
 }